unify swift-drive-autopilot config across clusters
This simplification is made possible by two changes:

1. The autopilot will not complain about an unmatching glob on
`/dev/sd[a-z][a-z]` when there are fewer than 26 drives.

2. The autopilot guarantees (both by documentation and by test coverage)
that swift-ids will always be assigned in order, so we can just put
`swift-01` through `swift-99` into the pool, and it will only use the
lowest-numbered IDs (e.g. `swift-01` through `swift-12` if there are 12
eligible drives).