increases resolve timeout to reduce flapping
We see alerts being resolved and fired again when a condition is flapping on and off. The resolution is triggered prematurely while the problem actually still persists. The next alert is fired soon after. Rinse repeat. Flapping alerts. Everyone annoyed. 

Most conditions and tests are executed every 5 minutes, like all the Blackox tests. With a resolution timeout of 5 minutes and alert will be resolved as soon as the next test succeeds. Setting this to 15 minutes requires that the test is executed succesfully 3 times in a row. Flaps should then occur less frequently.

1 minute bonus to account for scrap timeouts => 16 minutes