Prepare for statsd-exporter v0.5
statsd-exporter v0.5 removed the automatic suffixing of metrics based on
its kind: _timer, _gauge, _counter

Unfortunatelly that setting was switched on per defualz in 0.3 and all our
framework around that metrics are built on top: alerts and dashboards.

v0.5 Removes completly the support for that, so before updating the
exporter, we need to clean up. That should allow a better transition as we
cannot assume that we can update all regions at once.