Enables SSL Client Certificate Authentication
This PR enables authentication via SSL client certificates. It subs out
the basic auth with proxy auth. Grafana now expects the user to be
passed in via the `X-Remote-User` header. This can be toggled with the
`tls_client_auth.enabled` flag in the Helm values. It defaults to off
and keeps things as they are.

The ingress is configured to verify SSL certificates using the
`nginx.ingress.kubernetes.io/auth-tls-*` annotations. This can be turned
on/off per ingress spec.

Additionally, this PR injects a bit of logic into the ingress config, to
copy the relevant part (the user-id) into the `X-Remote-User` header.