Add interconnect alerts
The idea is to have critical alerts when a lot of regions can’t connect
and a warning if one region can’t connect so that we have less alert fatigue.
The critical alert is not yet set critical because I don’t want to let
it loose over the weekend.