add openstack_assignments_per_role metric to Keystone pgmetrics
I would like to monitor that restricted roles like resource_service and
swiftreseller do not get assigned to users and groups that I'm not aware
of. I thought about writing an exporter that queries the Keystone API
for role assignments, but realized that it's way more efficient to just
add a rule to the pgmetrics exporter of Keystone itself.