[nova] Disable io_ops weigher
Since there are multiple actual hosts behind one nova-compute service in
our setup, we don't need the io_ops weigher to have the same effect on
the weight as the RAM and disk weighers. But since we're not sure, what
a good weight is, we disable it for now. There can't be more than
currently 10 io_ops (read this as "instances in building/resizing/not
ACTIVE state") anyways, because the io_ops filter prohibits that.