[nova] soft-affinity has more weight
There's a "bug" in the nova-scheduler for soft-affinity, that makes it
return values < 1 for the host having most of the instances of a server
group. This happens, when a - it can be another one - server group has
had more instances deployed on a host, because of caching behaviour
implemented in
https://github.com/sapcc/nova/commit/c126d36640e0398e76ba01783b7f21f01f53a5f5
Effectively, this makes weight of a newly-created server group's second
instance 1 / max_server_group_members. For host with similar resources,
this can then make a big difference in the decision of the scheduler.

We now set the value to double the max_server_group_members value by
default, so it even matters against a newly-added host with lot of free
resources (which gains a +2 from ram+disk) against a nearly-full host
(which then gets at least 2 for affinity + a little ram+disk).