add context inhibition
Reintroduces the functionality to inhibit on `context`. Here we are guarding the inhibition of alerts without `context` by using a regex matcher for a non-empty `context`.

Additionally, I removed the `context` from `alertname` inihibition, as this didn't work. The point of `context` is to group alerts with different names. So, by definition that did nothing.

I also noticed that the `info` inhibition should not work at all. There's an regex for `severity` in a non-regex matcher.