[nova] Force resize to same host
We've added a patch to nova, that forces the host for a resize to be set
to the host the instance is on. The older patch we used for mitaka
doesn't work anymore. Probably because nova uses a request-spec now,
where our old force isn't added into. But it would be reseted anyways,
as our first approaches have shown.