Use yellow color for mixed notifications
This commit changes the color of the slack notification when it contains both firing and resolved alerts from read (danger) to yellow (warning).
In a lot of cases a notification that carries a "resolved" alert is an improvment: Some alerts have cleared while others are still firing. Marking them as yellow makes them easily distingushibale from alerts that only contain "firing" alerts which are never good.