[ironic] Ignore tempest-created resource-classes
The kos-operator would in any region running tempest create an entry to
nova's `quota_classes` table every time the ironic tempest tests run.
This slows down certain API calls quite a lot, so we ignore them.

We cannot use a variable in the list-comprehension because of how
contexts work in the kos-operator - it only searched globally for
variables, not locally.