[nova] Make dvs-agent inclusion configurable
With the switch to nsx-t, which is deployed by the neutron helm-chart,
we can get rid of the dvs-agent deployment in nova. Setting the variable
`neutron_dvs_agent_enabled` to `False` for a region, removes the
dvs-agent deployment. One might have to manuall delete the deployment of
the nova-compute pod though, because the vcenter-operator is not
supposed to delete stuff and most probably cannot update the deployment.