setup Maia scraping of Castellum such that it most likely works
I cannot get the service-discovery-based methods with the maia.io/scrape
annotations to work, and even if it worked, it probably would not yield
the expected results: Maia discards a whole lot of internal metrics on
ingestion, but those labels are necessary to differentiate the counter
metrics emitted by the various Castellum pods.

The only way I can imagine this working is by aggregating in the
prometheus-collector, then federating the aggregated metrics into Maia.
That's what this commit sets up.