[nova] Set regional domain for hostname in metadata
Cloud-init sets the hostname of a VM to whatever is set in the metadata.
There, the default is to add the domain `novalocal` to the hostname. We
didn't realize this, because usually the DHCP-client on the VM sets the
domain to `openstack.$region.cloud.sap` as provided by the
neutron-dhcp-agent. But when cloud-init is re-run, the domain is set to
`novalocal` again, so we better fix it.