[nova] Automate initial cellv2 setup
This wasn't necessary before, because we just recently changed to cellv2
and had to migrate the old setup. But for new regions and the new labs,
we need this automated.

This also refactors the script a bit to re-use common functionality and
extends the for-loop to also include cell into the `grep`. This also
means, that the cell0 `database_connection` will now be updated if it
changes as well. This was already the case for cell1 and cell2.

Additionally, this commit updates the `cell0_db_path`-helper, but it was
unused before, so nothing needs changing.