[nova] Add AggregateMultiTenancyIsolation to scheduler-filters
We want to be able to reserve BBs for certain customers. This can be
done on a project level by the nova-scheduler via the filter
`AggregateMultiTenancyIsolation` by adding a host to an aggregate with
the `filter_tenant_id` attribute set to a list of project-ids.

Every host that's not in any aggregate with that attribute passes this
filter. So nothing changes unless we add new aggregates.