[nova] Set per-HV reservation defaults
We later want to set different values for the hypervisors running HANA,
but for now need to reserve some head-space for the HV-software on every
ESXi host in nova. This does not ensure the space on every HV, but
rather makes nova not fill up the cluster too much with our limited
possibilities of managing the cluster from nova-side.
For more explanations, please have a look at the sapcc/nova repository's
commits regarding this feature.

5 % is the current minimum we got as recommendation from VMware
consultants.