[nova] Enable flavor-based HV-sizes for bigVMs
This makes the nova scheduler filter bigVMs map flavors to hypervisor
sizes based on an extra_specs attribute `host_fraction`.
We add two new host_fractions here to the default "full": one for "half"
a HV and one for "two_thirds" of a HV. Remeber that the filter also
allows flavors matching up to 10 % below that value.