[nova] Add new metric for last instance_action
When a nova-compute node hangs, e.g. because it was restarted and now
tries to power off a VM that has hanging tasks in the vCenter, we should
be able to see it. One way is to look at the time passed since the last
instance_action was created for that node. This metric will serve as
basis for an alert.