[nova] Increase max concurrent builds
The default in upstream nova is meant for single hosts hosting libvirt.
Our nova-compute runs against a cluster of hosts, where each host is
supposedly capable of running 8 builds in parallel. With 10 nodes in the
cluster, we still chose a conservative value of 50 here.

Since we increase max_concurrent_builds, we can also increase the number
of concurrent builds per project. This was bugging S4 a lot, as they
have a timeout on when their machines are supposed to be available and
they create a lot of machines at the same time in the same project.