[Nova] Increase prefer-same-host-on-resize multiplier
This weigher is designed to select the instance's current
host when given a chance. A bad combination of e.g. ram
available and disk space available might result in the host
not being rated first, although it should be (because the
filters have already decided that the host is eligible).

By setting the multiplier to 5 (which is equal to or more
than the other currently active weighers combined) the
prefer-same-host-on-resize weigher can overpower the others
if needed.