[nova] Set soft_affinity_weight_multiplier back to default
We've increased the weigt in aae05df381f3d303332ef15551f1cea9741090fa to
combat a bug in nova. With that bug being fixed upstream and us
backporting the commit, we are now able to remove this workaround. It
also didn't work properly anymore, because we started setting the
server_group_members via limes.