Simplifying PX alarming
PX alarming used to be hierarchical and had multible escalations, e.g. a single
BGP connection down not beeing as bad as all BGP connections to a consumer
router down. Never the less this concept was based on availability zones such
as eu-de-1 beeing treated as a single DC. This concept does not apply to
platforms such as HEC or GMP. Furthermore the multiple alarming escalations
seemed to complicate peoples' understanding, hence they are being removed for 2
escalations. DomainDown and BGPDown.