[nova] Create snapshots from full clones
With the current way we create our instances, they're mostly eagerly
zeroed, because the CopyVirtualDisk_Task always converts files to
eagerly zeroed. This can lead to problems - especially on Windows VMs -
where the ESXi doesn't send any data for a longer period of time while
reading and compressing all the zeroes on export. Since nova-compute
acts as a proxy to glance, which again acts as a proxy to swift, the
HTTP request to swift times out with a timeout, if it receives no data
for a certain period of time.

We can mitigate this problem by converting the volume to either thin or
lazy-zeroed, because ESXi can then skip the blocks. The conversion to
the latter happens automatically if we do a full clone before exporting
the VM.  Therefore, we enable the `full_clone_snapshots` option.