[nova] Always start metadata service in own pod
When having multiple cells, we disabled the standalone metadata service,
because that's not supported. Instead, one has to run the metadata
service "globally" as an api service. But this only means, one should
not use the "nova-api-metadata" executable to run the metadata service.

Therefore, we now run the metadata service in its own pod with the
nova-api executable.

We need to make sure we switch neutron to using the standalone api
endpoint everywhere before completely removing metadata from the api
service.

Afterwards, we should be able to switch both api and metadata to using
uwsgi. This is not possible with nova-api also having to spawn the
metadata service on a different port.