[nova] Resolve global.novaApi.executable and debug mode
With changing the secrets merging strategy to requiring an approved
pull-request, it doesn't make sense to enable debug mode via secrets
change in a region - it just takes too long. Additionally, I can't
remember a single case where we would have needed it.

In that process, we also resolve the global.novaApi.executable setting,
because I don't see any sense in changing this via secrets and it's only
used in a single place, so there's no need for a variable.