[cinder] Add initial alert based on openstack-exporter
This patch adds a simple info alert for now for cinder.
The alert watches the amount of overcommit ratio left is
less than 0.5.

The amount of overcommit left is the difference between
the cinder configured max over subscription ratio and the
calculated overcommit ratio from allocated storage.

Typically the max oversubscription ratio is set to 1.5.