[nova] Add openstack_compute_nodes_* metrics again
We might not have the AZ in there, because we can't query it from the
DB, but at least we have some node infos. Once we find out, if and where
we need the AZ, we need to add another pgmetrics pod querying the
nova_api DB to add a metric we can use to assign compute nodes to AZs.