keppel: restrict number of database connections
In prod, I'm usually seeing 15-20 connections during normal operation,
so this limit should be plenty. If it *is* reached, we might get some
more informative errors. Right now, we're only seeing the postgres die
to OOM, which *might* be caused by a client connection storm, but I
don't know for sure.