[sre] Define service by first part of ingress
Most of our ingresses have the actual service's name as first part of
the ingress-name, e.g. ironic-api, neutron-server,
designate-mariadb-backup, limes-api-ccloud. Since we're not allowed to
change the regex on the prometheus side to route everything beginning
with a "$service-" to the appropriate slack channel, because the
responsible person sees this as a too complicated change to the regex,
we're force to do it on our alerts instead - even though it does make
much more sense to do it centrally.