[nova] Prohibit missing openstack_compute_stuck_instances_count_gauge
When there are no instances in the filtered states, the SQL query
providing this metric returns no rows. Since we introduced an operator
checking for absent metrics used in an alert, we now get alerts for the
previously described condition. We don't want that, so we always query
out a dummy row which should not trigger any alert as its values do not
match any alert filter.