[cinder] Prohibit missing openstack_stuck_volumes_max_duration_gauge
When there are not volumes stuck in attaching/deleting or detaching
state for a longer period of time, the query backing
openstack_stuck_volumes_max_duration_gauge will not return any result
making the metric disappear. This makes the absent metrics operator
alert us. We don't want that, so we always query out a dummy row which
should not trigger any alert as its values do not match any alert
filter.