[nova] Spread api workers over more pods
We used to have the default number of workers (number of CPUs) for
nova-api, but that doesn't scale well, makes for big resource requests
that are hard to schedule  and also overloads our kubernetes nodes at
times. Therefore, we now use only 8 workers per pod and scale nova-api
up to 3 replicas for a total of 24 workers. This should fit the number
of sql connection possible to pgbouncer again (450 in total per db -> 24
* 15 = 360) leaving some head-room for scaling up, pod restarts and
* nova-api-metadata.