[nova] Spawn more vspc pods with better resource-requests
We increase the number of pods, because we only have a single process
doing its stuff with asyncio. This can be very limiting at times.

This commit includes a change to let nova-vspc use the same dictionary
as other pods for defining the replicas.

We can reduce the amount of requested memory, because we use more pods
and spread the load. But still, we need to request more CPU to handle
it, as we've seen constant CPU-load above 15 % in some regions.