[cinder] Optionally run cinder-api via uWSGI
By setting "api.use_uwgi = true" in a region, we can run cinder-api via
uWSGI instead of the eventlet WSGI server. This can be more performant,
but when choosing the number of workers, we have to bear in mind, that
there are no greenthreads in them - in the eventlet server, there were
100 greenthreads in each process.

This also means, that the number of DB connections required per worker
should decrease significantly - in theory to the number of workers, but
we've seen manila-api use 2 connections per worker, so this needs to be
tested.