[nova] Support adding compute-nodes to shard-cells
In qa-de-1 we've created a cell based only on a single shard instead of
a whole availability zone. For this, we need to update the template to
detect other cells. We prioritize availability zone cells over shard
cells, because that's how we started out. If we ever start creating more
cells, we might want to change the order.