[nova] Increase http_retries for neutron
We've incresed the http_retries for cinder, which seemed to have solved
most if not all cases of `ConnectFailure` to cinder being raised,
according to a view through all region's sentry. The only on thing still
popping up is `ConnectFailure` for Neutron. Therefore, we also increase
the http_retries from 3 to 11, making us resilient against Neutron being
unconnectable for 5s instead of < 2s.