[nova] Switch to sentry-operator
Instead of manually updating/adding the "sentry_dsn" key in/to the
secrets after adding its project to sentry, we now tell the
sentry-operator to add the project and provide the DSN to us via
k8s secret.

We also add the `SENTRY_DSN` environment variable to all services, that
didn't have it before.