[nova] Add a timeout to requests to Neutron
We've started to see problems recently in na-us-2, where building a VM
would hang on a request to Neutron for updating the ports. This keeps
the lock of that VM for multiple hours and every later issued delete
call would wait on the lock, thus exhausting the number of greenthreads.
This leads to messages piling up in rabbitmq, as nova-compute cannot
work on them anymore.