[osprofiler] clean up helper functions (#2231)
This PR clears old osprofiler setup with flamegraph-redis.

Also in the old setup, both global value global.osprofiler.enabled and service local osprofiler.enabled are checked in the helper function definition. It is bit confusing to control the integration of osprofiler, since osporfiler is not enabled for all services and one has to edit different places. With this PR, osprofiler integration can only be enabled explicitly by the service local value osprofiler.enabled.

Co-authored-by: Chuan Miao <chuan.miao@sap.com>