[nova] Require explicit password for nova user
We did set an explicit password in most regions, but it might happen,
that we forget about it and thus generate one automatically. This will
not work if anybody else is relying on the password being set.
Additionally, we need to switch away from one global master_password,
to service-based master passwords, which takes away most of the
advantage of the master-password-algorithm.