nova: fire OpenstackNovaInstanceStuckSpawning less often
We see this alert firing time and time again in eu-de-1, where it's
usually an image-download that takes longer. The alerts resolve after
ca. 1h. While we would like to increase the alerting time only in
eu-de-1, there doesn't seem to be an easy way to achieve this. Thus, we
increase the time in all regions.