[manila] increase migration access rule sync timeout
For shares with hundreds of access rules this may indeed take a lot of time
and waiting does no harm, recovering after the migration timeout occurred is harder.
That would require to set the replica state for the active replica in the db,
reset the states and task_state of the share instances and remove the source
share instance.