[swift-proxy] enable 'delete_container_retry_count' for bulk operations.
Large Glance images may include 100s of segments per container.
Deletion of such images often fails with 'HTTP 409 Conflict' response
from Swift. Enabling this option seem to resolve the problem.

The delete_container_retry_count parameter is used during a bulk delete of objects
and their container. This would frequently fail because it is very likely that
all replicated objects have not been deleted by the time the middleware got a
successful response. It can be configured the number of retries. And the
number of seconds to wait between each retry will be 1.5**retry