[nova] Do not reserve 5% of vCPUs on ESXi
We don't need to do that, as ESXi will always take what it needs - other
than with RAM, where it would have to swap out VM pages. HANA on VMware
best practices [1] suggest to use all cores + HT cores of a NUMA node,
if the VM needs that NUMA node. For this to work, we cannot reserve 5 %
of vCPUs anymore.

[1] https://core.vmware.com/resource/sap-hana-hyperconverged-infrastructure-hci-solutions-powered-vmware-vsan#sec9-sub1