[ELK] Parse nginx ingress access logs and global-request-id
The previous version was buggy due to a '1' at the end of the request path,
so it is rather unlikely that we rely on the nameing of the values here.

The new version parses all the fields defined in the default access-logs
as they are named in the ingress nginx definition.

Additionally the global_request_id will be parsed, if set.
It strips the prefix "req-" in case it is there, so we can search just for the uuid,
and ensure that it matches the openstack logs, where we strip the "greq-" prefix.