[nova] Add cloud_volume_admin role to nova service user
We need this to talk to Cinder, as we have to configure the [cinder]
section so Nova can update volumes properly if there's not user-token
involved - e.g. when deleting an instance on nova-compute startup.

We're not using the cinder user to talk to Cinder, because changing the
password of cinder would require both Nova and Cinder to roll out at the
same time to have minimal downtime.