[nova] Seed role swiftoperator nova user in master project
We did not use this role before and thus need to seed it to keystone, if
it does not exist, yet. This can be the case, because on new buildups
Nova might be one of the earlier services to be deployed - even before
swift.

We need the nova@Default user to have swiftoperator in ccadmin/master
because we share images through this project and we want to switch Nova
to use the nova user to access Ironic. The ironic user already has the
swiftoperator role in ccadmin/master.