[neutron] Fix label-generation for f5 and asr1k agent
In the context of the functions defined in the templates, "." is not the
full environment context anymore, but the functions parameter instead -
at least the code looks like that. That's why we have to pass the
variable containing the full context explicitly instead of "." to the
snippet generating the labels.