[nova] Don't snapshot a VM before cloning its disk
There are multiple problems we want to fix by not snapshotting a VM
before cloning its root disk.
1) Snapshotting with volumes attached can fail, because there's not
enough space for the full snapshot on the vVOL backend.
2) Snapshotting an instance can fail, if there's an old/buggy
VMware-tools version running inside a Windows guest.
3) Snapshotting can fail, if one of the volumes is on a storage having
IO profiles.

Most of the problems stem from the fact, that we attach volumes in
dependent mode, which means they are snapshotted together with the
instance.

Since Cinder only does a clone without snapshot and we haven't seen
problems with that, we will switch Nova to do the same.