[nova] Disable KVM hypervisor templates
We currently don't have the capacity to run any KVM hypervisors and thus
don't have to include the templates into our charts. It also hurts us,
as helm-diff finds changes for images, which the image-cache in the CI
tries to cache, but that are no longer updated/built and thus caching
never completes.

This commit also enables finding images for the CI entrypoint in the
kos-template for KVM hypervisors via the isImageTransportTemplating
value.