[nova] Prepare Jobs for helm3
With Helm3, we don't force the replacements of Job specs anymore, which
causes deployment issues since kuberentes Job specs are immutable by default.

We solve this by generating dynamic Job names for every deployment. This
creates a new Job and the old one will be deleted.

For the migration Jobs, we create the name based on the image. The
assumption is, that we only need to run the migrations once and new
migrations get added in new images only.

We have to run the cell update on every release, because the
configuration might have changed, and thus create the name
from the automatically-incremented release number.