Remove default configuration from maintenance-controller
Certain values of the maintenance controller config support golang
templating. These have been piped through the 'tpl' function. So they
needed to be quoted. Golang templates use '{{}}', which could be JSON
objects for YAML and therefore require quoting. Defining a custom config
required multiple levels of quotation, which is a mess for humans as for
machines.