Move address-scope config into a seperate file
Address Scopes in the past were replicated in each of the asr1k agent
config sections and also in the neutron-aci config. They must be equal
in order for things to work. Hence it makes much more sense to move it
out to a separate file and render it in the relevant configmaps.