Add custom flavor for tests: hana_cc3test
This flavor is used by the cc3test project to check HANA-only BBs are
capable of spawning new VMs. It's private and needs to be shared to
project, as we don't want any customer to use it. We need it, because
we don't want to waste so many GiB of RAM and quota just for testing
spawning and volume-attachment.