system/gatekeeper: improve violation msg for GkPodLabels
Not using `required_labels` here as it is a set and we have to use
`json.marshal()` to get array in string to be consistent with other
messages. Marshaling results in array elements to not have any spaces in
between.

In other words, put spaces between missing labels in the message.