[Cinder] Set an overall reserved_percentage to 20%
We want to have cinder enforce that 20% of the overall storage
on a backend is reserved.  This means that cinder will only allow
allocating up to 80% of the total storage on a backend.  This will
be more meaningful when datastores as pools is rolled out as then
each datastore can only be allocated to a max of 80% capacity.