Nova: Derive name from job-script hash
Scripts can change partly due to code change or configuration
values (cells), which would result in an annotation change.
But k8s doesn't do updates of jobs, so we end up with a deployment
failure.

By deriving the name from the hashed script (instead of the config)
we run the job newly on each change of the script or the image version.