[Cinder] Add the resource entry to log context
This patch adds the %(resource)s to the log context in each cinder log
entry.  Various places in cinder debug log entries adds the
resource=volume_ref attribute.  For example:
https://github.com/sapcc/cinder/blob/stable/train-m3/cinder/volume/manager.py#L332-L333

https://github.com/sapcc/cinder/blob/stable/train-m3/cinder/volume/drivers/vmware/vmdk.py#L1201-L1202
https://github.com/sapcc/cinder/blob/stable/train-m3/cinder/volume/drivers/vmware/vmdk.py#L2900

This will result in the volume id being added to the log entry whenever
the LOG.debug() has the kwarg of resource=volume.  This will help
with tracking down what volume was being operated on in the logs.