Rabbitmq: Option for statefulSet without clustering
By default, this change doesn't change anything, except it
will raise an error when a deployment with more than one
replica with persistence enabled is requested, as this doesn't work.

One has to use a statefulset for that, but it is *not* a clustered
setup. The use-case is a special one for notifications:

We currently create for each service a single independent
rabbitmq, and the consumer connects to each of them.

If one goes down, no notifications can be send (over rabbitmq).

This allows us to spin up a statefulset with several independent
on the consumer side, let the producers connect to any of them,
and the consumer to all of them.

When one instance is down, the producer(s) can reconnect to
any of the remaining ones.
A Pod disruption budget ensures that at least one is up.

Clustering is not needed, as we do not need a consistent
state within rabbitmq, as we didn't have that before either.
We only need to ensure, that all messages put on the queues
will be delivered to the consumer.