[openstack/utils] Add pre_stop_graceful_shutdown snippet
Add a common way to handle the propagation delay of a pod termination
gracefully.
In order to get around the race between the kubernetes stop sending
new connections to a pod, and the pod accepting them,
add a delay to the shutdown of the pod.
Ideally, this would be replaced with some insight into the actual
state of forwarding, but until then we have to wait