[nova] Optionally enable RPC metrics collection
We've added a patch to oslo.messaging, that allows us to track the
number of RPC calls per called function and also if this call errored
out. This functionality can be enabled by setting statsd_enabled in the
DEFAULT section.

This change adds the possibility to enable this functionality for
nova-scheduler, nova-conductor and nova-compute pods (both ironic and
vmware).