Increasing VMX-HW Version to 18 for all flavors
We set the default VMX-HW Version in Nova for new VMs to 18. We do not change the VMX-HW Version if a VM gets resized in Nova, except the VMs which get resized from a Flavor with < 128 vCPUs to a Flavor with > 128 vCPUs. For them the resize will change the VMX-HW version to the defined one in the flavor because otherwise the VM would not start up.