[nova] Update default_hw_version to vmx-18
We introduced the default_hw_version to define a minimum version
supported by all ESXi during upgrade, so that vMotion between ESXis and
clusters is possible no matter if they are upgraded already or not. The
minimal supported version at the time was vmx-13. We now upgraded to
newer versions of ESXi and thus update default_hw_version to what the
ESXis would use as default currently: vmx-18.