[nova] Disable RamFilter and DiskFilter in scheduler
With filtering happing in placement, it's already ensured that the
resources in the flavor - MEMORY_MB and DISK_GB in placement - can be
fulfilled by the host. Therefore, we don't need these filters in the
scheduler anymore.

This came up because baremetal flavors use resource group based
scheduling and set MEMORY_MB, VCPU and DISK_GB to 0 in the request to
placement via flavor property. Still, a baremetal host could not be
scheduled because of 1 GiB difference between placement and the flavor's
disk setting.