Use a k8s job for config reload
Instead of relying on the pipeline to reload the configuration, we can
also do this by a k8s job.
This is is a draft, we are missing the feature to wait that the new
configmap has been distributed and we also need to make sure that the
pod we are referring to is running.