[nova] Use single statsd-exporter.yaml
We recently introduced an additional statsd-rpc-exporter.yaml, which did
not include the default settings, but only the mappings for
oslo.messaging.* metrics. When introducing the new config file, the
vmware hypervisor pods did not get an update.

Since we want both the default settings and the mappings in all statsd
exporters we run, we now use a single file for both in all the pods.