[nova] Switch for standalone placement
This splits out the placement config, and adds an option
to disable deploying placement from this chart.

To migrate, deploy first the standalone placement chart
(different changeset).
This will update the endpoint in keystone, so all request
should be routed through the new deployment.

Then deploy this chart with placement.enabled=False,
which will undeploy the nova internal placement.

The dependency of the data in the nova_api database needs
to be solved separately.