[swift] Avoid multiple scrapes of prometheus targets
The service discovery is invoked for every endpoint port found in the correponding endpoints object and in both cases the port annotation is used to override the port. This will cause the same metrics to be collected twice with different container labels.

Solution: Remove the prometheus.io/port annotation from your pod or service spec and give the target port the name metrics.