[mariadb] Pre-upgrade hook for db/user creation
The idea is to roll-out new users with new credentials in a pre-upgrade hook
so that they are available to the actual deployment making use of it.

Users (or databases) cannot be deleted this way (much less rolled back).
The cleanup is left to a later iteration, which may make use of an expiry policy.

Instead of having custom init scripts with slightly different patterns,
the fields `databases` and `users` are there to create and update them.

This also allows for uniform access to other macros, such as side-car
containers, which need knowledge about the users and the databases as well.

By using a job which executes the change within the container
of the database, we can change the credentials of any user
without fearing a lockout.

The job will be executed on upgrade and rollback and since it
gets the values from the previous deployment, it not only
will set the new credentials on upgrade, it also restores the old ones
on rollback.